requirements.txt
google-api-python-client==1.2
python-dateutil==2.2
python-gflags==2.0
